% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matrianalyze}
\alias{matrianalyze}
\title{matrianalyze: tabulations of matriannotated data}
\usage{
matrianalyze(data = NULL)
}
\arguments{
\item{data}{character. The matriannotated input data.}
}
\value{
a tabulated data.frame.
}
\description{
matrianalyze: tabulations of matriannotated data
}
\examples{
matriannotate(data)
}
