% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matriannotate}
\alias{matriannotate}
\title{matriannotate: a function to add matrisome annotations to gene tables}
\usage{
matriannotate(data = NULL, gene.column = NULL, species = NULL)
}
\arguments{
\item{data}{character. The input data.}

\item{gene.column}{character. The name of the input column with gene names (accepted formats are Gene Symbols, NCBI gene ID (formerly Entrez Gene ID) and Ensembl Gene ID).}

\item{species}{character. One of the species for which matrisome annotations are provided (human, mouse, c.elegans, drosophila, zebrafish).}
}
\value{
The input data with added matrisome categories and families.
}
\description{
matriannotate: a function to add matrisome annotations to gene tables
}
\examples{
matriannotate(data, gene.column, species)
}
