% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matriflow}
\alias{matriflow}
\title{matriflow: alluvials for matriannotated data}
\usage{
matriflow(data = NULL, print.plot = TRUE)
}
\arguments{
\item{data}{character. The matriannotated input data.}

\item{print.plot}{logical. If TRUE (default) the plot is printed, otherwise the graph object (in its original format) is returned.}
}
\value{
a ggplot2 object (printed or not)
}
\description{
matriflow: alluvials for matriannotated data
}
\examples{
matriflow(data)
}
