% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matripie}
\alias{matripie}
\title{matripie: donutpie for matriannotated data}
\usage{
matripie(data = NULL)
}
\arguments{
\item{data}{character. The matriannotated input data.}
}
\value{
a DonutPie object (always printed to screen)
}
\description{
matripie: donutpie for matriannotated data
}
\examples{
matripie(data)
}
